# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "APIServer",
    "version": "v1"
  },
  "paths": {
...
    "/Posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
...
    "/Posts/{id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Songs": {
      "get": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Songs"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          }
        }
      }
    },
...
    "/Songs/Genre/{option}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "option",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Songs/{id}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Song": {
        "required": [
          "album",
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "album": {
            "$ref": "#/components/schemas/Album"
          },
          "title": {
            "type": "string"
          },
          "genre": {
            "type": "string",
            "nullable": true
          },
          "soundCloud": {
            "type": "string",
            "nullable": true
          },
          "audio": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Album": {
        "required": [
          "genre",
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            },
            "nullable": true
          },
          "coverImage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Post": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Media"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MediaType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "Media": {
        "required": [
          "type",
          "url"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          },
          "article": {
            "$ref": "#/components/schemas/Article"
          },
          "url": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/MediaType"
          }
        },
        "additionalProperties": false
      },
      "Article": {
        "required": [
          "content",
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Media"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}